+++ src/main/kotlin/com/tul/shoppingcart/config/ShoppingCartConfig.kt
@@ -1,0 +1,9 @@
+package com.tul.shoppingcart.config
+
+import org.springframework.context.annotation.Configuration
+
+
+@Configuration
+class ShoppingCartConfig {
+
+}