+++ src/main/kotlin/com/tul/shoppingcart/dto/CartDTO.kt
@@ -1,0 +1,8 @@
+package com.tul.shoppingcart.dto
+
+import com.tul.shoppingcart.enums.StatusEnum
+
+class CartDTO (
+    val products: HashMap<ProductDTO, Int>,
+    val status: StatusEnum
+)