+++ src/main/kotlin/com/tul/shoppingcart/enums/StatusEnum.kt
@@ -1,0 +1,6 @@
+package com.tul.shoppingcart.enums
+
+enum class StatusEnum (val value: String, description: String) {
+    PENDING("PENDING", "El carrito se encuentra pendiente"),
+    COMPLETED("COMPLETED", "El carrito fue completado")
+}