+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 
 This file will show the step by step of the commits since project's creation
 
-## 17/01/2022 13:40
+## 17/01/2022 15:40 - Branch: feathre/h2-database
+- H2 database to use as a repository of products and carts.
+  
+## 17/01/2022 13:40 - Branch: feature/controller_first_approach
 - Adding cart controller and products controller
   
-## 17/01/2022 13:00
+## 17/01/2022 13:00 - Branch: feature/controller_first_approach
 - Controller first approach
 ![img.png](img.png)